requirements.txt
requests==2.32.3
beautifulsoup4==4.12.3
python-dotenv==1.0.1 
transformers==4.42.3
torch == 2.3.1
google-ai-generativelanguage==0.6.6
selenium == 4.22.0
